Production Build for NextJS Fixed
