+++ frontend/next.config.mjs
@@ -6,6 +6,7 @@ const nextConfig = {
     images: {
         domains: ['images.unsplash.com', 'd9amksc9hkkpj.cloudfront.net'], // Add this line to configure the domain
       },
+    output: 'standalone'
 };
 
 export default nextConfig;
