+++ frontend/package.json
@@ -35,7 +35,8 @@
     "next": "15.1.6",
     "react": "^19.0.0",
     "react-color": "^2.19.3",
-    "react-dom": "^19.0.0"
+    "react-dom": "^19.0.0",
+    "sharp": "^0.34.1"
   },
   "devDependencies": {
     "@babel/core": "^7.26.9",
