+++ frontend/src/app/Profile/page.jsx
@@ -14,7 +14,7 @@ import ActivityGrid from "../../components/ActivityGrid";
 import Navbar from "../../components/Navbar";
 import ProfilePicture from "../../components/ProfilePicture";
 
-export default function ProfilePage({ children }) {
+export default function ProfilePage() {
     const [image, setImage] = useState();
     const [profileData, setProfileData] = useState({});
     const [errorMessage, setErrorMessage] = useState("");
