--- frontend/src/app/blogs/page.jsx
@@ -1,12 +1,0 @@
-import { Room } from "../../components/Room";
-import { Editor } from "../../components/Editor";
-import React from "react";
-
-export default function Page() {
-  return (
-    // <Room>
-    //   <Editor/>
-    // </Room>
-    hi
-  );
-}
